Big changes with login logout register with flask jwt

+++ resources/item.py
@@ -1,6 +1,7 @@
 from flask import request
+from flask_jwt_extended.utils import get_jwt_identity
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 # items = []
@@ -24,6 +25,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not Found!'}, 404
     
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': f'An Item with name {name} already exists'}, 400
@@ -35,7 +37,11 @@ class Item(Resource):
             return {'message': 'An Error occured inserting the item'}, 500
         return new_item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin priilege Required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -56,6 +62,13 @@ class Item(Resource):
 
     
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'Items': [x.json() for x in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        if user_id:
+            return {'Items': [item.json() for item in ItemModel.query.all()]}, 200
+        return {
+            'Items': [item.name for item in ItemModel.query.all()],
+            'message': 'More Data available if you login' 
+            }, 200
